+++ pytorch_lightning/trainer/deprecated_api.py
@@ -137,7 +137,6 @@ class DeprecatedTrainerAttributes:
     accelerator: Accelerator
     lightning_module = LightningModule
     sanity_checking: bool
-    evaluated_ckpt_path: str
 
     @property
     def accelerator_backend(self) -> Accelerator:
@@ -161,11 +160,3 @@ class DeprecatedTrainerAttributes:
             " and will be removed in v1.5.", DeprecationWarning
         )
         return self.sanity_checking
-
-    @property
-    def tested_ckpt_path(self):
-        rank_zero_warn(
-            '`Trainer.tested_ckpt_path` has been renamed to `Trainer.evaluated_ckpt_path`'
-            ' and will be removed in v1.5.', DeprecationWarning
-        )
-        return self.evaluated_ckpt_path